Fix bug in manage page with self-selection.
Participation links are internally linked to the cohort that was self selected by contributors.
Make sure this link is not revealed to the analyst when checking the history or retrieving previous generated participation links.

Solution: if cohort self selection is on, always override cohort id with undefined (for no cohort) in server response to get_client_urls. Otherwise (no cohorts, or pre-assigned) keep cohort mapping as it is.